refactor interfaces in to ingress package
